+++ plugins/github-deployments/src/api/index.ts
@@ -58,6 +58,10 @@ export type GithubDeployment = {
     abbreviatedOid: string;
     commitUrl: string;
   };
+  creator: {
+    login: string;
+  };
+  payload: string;
 };
 
 type QueryParams = {
@@ -99,6 +103,10 @@ query deployments($owner: String!, $repo: String!, $last: Int) {
           abbreviatedOid
           commitUrl
         }
+        creator {
+          login
+        }
+        payload
       }
     }
   }