+++ plugins/github-deployments/src/components/GithubDeploymentsCard.test.tsx
@@ -26,7 +26,11 @@ import {
 
 import { fireEvent } from '@testing-library/react';
 import { msw, renderInTestApp } from '@backstage/test-utils';
-import { GithubDeploymentsApiClient, githubDeploymentsApiRef } from '../api';
+import {
+  GithubDeployment,
+  GithubDeploymentsApiClient,
+  githubDeploymentsApiRef,
+} from '../api';
 import { githubDeploymentsPlugin } from '../plugin';
 import { GithubDeploymentsCard } from './GithubDeploymentsCard';
 
@@ -41,6 +45,8 @@ import { setupServer } from 'msw/node';
 import { graphql } from 'msw';
 import { ScmIntegrations } from '@backstage/integration';
 import { Entity } from '@backstage/catalog-model';
+import { GithubDeploymentsTable } from './GithubDeploymentsTable';
+import { Box } from '@material-ui/core';
 
 let entity: { entity: Entity };
 
@@ -203,6 +209,40 @@ describe('github-deployments', () => {
         await rendered.findByText('GitHub Deployments'),
       ).toBeInTheDocument();
       expect(await rendered.findByText('failure')).toBeInTheDocument();
+    });
+    
+    it('should display extra columns', async () => {
+      worker.use(
+        graphql.query('deployments', (_, res, ctx) =>
+          res(ctx.data(responseStub)),
+        ),
+      );
+
+      const renderTargetFromPayload = (payload: string) => {
+        const parsedPayload = JSON.parse(payload);
+        return parsedPayload?.target || 'unknown';
+      };
+
+      const extraColumn = {
+        title: 'Target',
+        render: (row: GithubDeployment): JSX.Element => (
+          <Box>{renderTargetFromPayload(row.payload)}</Box>
+        ),
+      };
+
+      const columns = [
+        ...GithubDeploymentsTable.defaultDeploymentColumns,
+        extraColumn,
+      ];
+
+      const rendered = await renderInTestApp(
+        <ApiProvider apis={apis}>
+          <GithubDeploymentsCard columns={columns} />
+        </ApiProvider>,
+      );
+
+      expect(await rendered.findByText('moon')).toBeInTheDocument();
+      expect(await rendered.findByText('sun')).toBeInTheDocument();
     });
 
     describe('entity with source location', () => {