+++ plugins/github-deployments/src/mocks/mocks.ts
@@ -48,6 +48,10 @@ export const responseStub: QueryResponse = {
             commitUrl: 'https://exampleapi.com/123456789',
             abbreviatedOid: '12345',
           },
+          creator: {
+            login: 'robot-user-001',
+          },
+          payload: '{"target":"moon"}',
         },
         {
           state: 'pending',
@@ -57,6 +61,10 @@ export const responseStub: QueryResponse = {
             commitUrl: 'https://exampleapi.com/543212345',
             abbreviatedOid: '54321',
           },
+          creator: {
+            login: 'robot-user-002',
+          },
+          payload: '{"target":"sun"}',
         },
       ],
     },
@@ -75,6 +83,10 @@ export const refreshedResponseStub: QueryResponse = {
             commitUrl: 'https://exampleapi.com/123456789',
             abbreviatedOid: '12345',
           },
+          creator: {
+            login: 'robot-user-001',
+          },
+          payload: '',
         },
         {
           state: 'failure',
@@ -84,6 +96,10 @@ export const refreshedResponseStub: QueryResponse = {
             commitUrl: 'https://exampleapi.com/543212345',
             abbreviatedOid: '54321',
           },
+          creator: {
+            login: 'robot-user-002',
+          },
+          payload: '',
         },
       ],
     },