+++ plugins/techdocs-backend/src/service/router.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
-import { Entity } from '@backstage/catalog-model';
+import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 import { Config } from '@backstage/config';
+import { NotFoundError } from '@backstage/errors';
 import {
   GeneratorBuilder,
   getLocationForEntity,
@@ -30,8 +31,6 @@ import { Knex } from 'knex';
 import { Logger } from 'winston';
 import { DocsBuilder } from '../DocsBuilder';
 import { shouldCheckForUpdate } from '../DocsBuilder/BuildMetadataStorage';
-import { getEntityNameFromUrlPath } from './helpers';
-import { NotFoundError } from '@backstage/errors';
 
 type RouterOptions = {
   preparers: PreparerBuilder;
@@ -55,10 +54,9 @@ export async function createRouter({
 }: RouterOptions): Promise<express.Router> {
   const router = Router();
 
-  router.get('/metadata/techdocs/*', async (req, res) => {
-    // path is `:namespace/:kind:/:name`
-    const { '0': path } = req.params;
-    const entityName = getEntityNameFromUrlPath(path);
+  router.get('/metadata/techdocs/:namespace/:kind/:name', async (req, res) => {
+    const { kind, namespace, name } = req.params;
+    const entityName = { kind, namespace, name };
 
     try {
       const techdocsMetadata = await publisher.fetchTechDocsMetadata(
@@ -67,14 +65,15 @@ export async function createRouter({
 
       res.json(techdocsMetadata);
     } catch (err) {
-      logger.error(
-        `Unable to get metadata for ${entityName.namespace}/${entityName.name} with error ${err}`,
+      logger.info(
+        `Unable to get metadata for '${stringifyEntityRef(
+          entityName,
+        )}' with error ${err}`,
       );
-      res
-        .status(500)
-        .send(
-          `Unable to get metadata for $${entityName.namespace}/${entityName.name}, reason: ${err}`,
-        );
+      throw new NotFoundError(
+        `Unable to get metadata for '${stringifyEntityRef(entityName)}'`,
+        err,
+      );
     }
   });
 
@@ -82,9 +81,11 @@ export async function createRouter({
     const catalogUrl = await discovery.getBaseUrl('catalog');
 
     const { kind, namespace, name } = req.params;
+    const entityName = { kind, namespace, name };
 
     try {
       const token = getBearerToken(req.headers.authorization);
+      // TODO: Consider using the catalog client here
       const entity = (await (
         await fetch(
           `${catalogUrl}/entities/by-name/${kind}/${namespace}/${name}`,
@@ -98,10 +99,13 @@ export async function createRouter({
       res.json({ ...entity, locationMetadata });
     } catch (err) {
       logger.info(
-        `Unable to get metadata for ${kind}/${namespace}/${name} with error ${err}`,
+        `Unable to get metadata for '${stringifyEntityRef(
+          entityName,
+        )}' with error ${err}`,
       );
-      throw new Error(
-        `Unable to get metadata for ${kind}/${namespace}/${name} with error ${err}`,
+      throw new NotFoundError(
+        `Unable to get metadata for '${stringifyEntityRef(entityName)}'`,
+        err,
       );
     }
   });